Create README.md
Text to Image Generation using Diffusion

For our Text-to-Image generation project we useddiffusion-based image generation UNet2DConditionModel of Hugging Face diffusers that can generate images directly from text prompts, while also visualizing the transformation of noisy latents into a coherent image through diffusion steps. 
We first tried training a custom PyTorch U-Net on CIFAR-10 at its native 32×32 resolution early in the project. Because this version required us to implement every part manually of the diffusion pipeline, including noise scheduling, conditioning logic, and training loops, it was harder for the model to tune and it is more computationally heavy, and it’s complex to implement. So, we decided to switch to using Hugging Face's production-ready diffusion components. By adopting the UNet2DConditionModel with in-built cross-attention, the DDPM scheduler, and CLIP's text encoder and tokenizer for prompt conditioning, we were able to simplify the architecture and improve model stability; train faster on Colab GPUs; and adopt a setup which closely resembles real-world text-to-image setups like Stable Diffusion-scaled down to 64×64 resolution for efficiency. 

We trained the model for 8 epochs on CIFAR-10 datasets where the image size we took is 64*64 for slightly improved clarity of images. 
So when we are training the model what we observe is the model learned to predict noise given both noisy images and text embeddings and average loss stabilized around 0.010–0.012, for good convergence for a model. 
We then generated a GIF containing 10 frames that shows the evolution of the image as noise gradually transforms into a recognizable CIFAR-style “dog”.